+++ main.go
@@ -5,17 +5,15 @@ import (
 	"html/template"
 	"net/http"
 	"strconv"
-	"strings"
 
+	"github.com/utkarsh-1905/thapar-time-table/utils"
 	"github.com/xuri/excelize/v2"
 )
 
 func main() {
 	f, err := excelize.OpenFile("timetable.xlsx")
 
-	if err != nil {
-		panic(err)
-	}
+	utils.HandleError(err)
 
 	defer func() {
 		if err = f.Close(); err != nil {
@@ -32,9 +30,11 @@ func main() {
 		Classes map[string]map[int]string
 	}
 
-	var sheets []string
+	sheets := f.GetSheetList()
+
 	classes := make(map[string]map[int]string)
-	sheets = f.GetSheetList()
+
+	//finding all classes in a sheet
 	for _, sheet := range sheets {
 		temp := make(map[int]string)
 		cols, err := f.GetRows(sheet)
@@ -48,9 +48,7 @@ func main() {
 			}
 		}
 		classes[sheet] = temp
-		if err != nil {
-			panic(err)
-		}
+		utils.HandleError(err)
 	}
 
 	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
@@ -85,79 +83,12 @@ func main() {
 			errorPage.Execute(w, "Invalid category/class combination")
 			return
 		}
-		table.Execute(w, GetTableData(sheet, class, f))
+		table.Execute(w, utils.GetTableData(sheet, class, f))
 	})
 
 	fmt.Println("Starting server at http://localhost:3000")
 	err = http.ListenAndServe(":3000", nil)
-	if err != nil {
-		panic(err)
-	}
+	utils.HandleError(err)
 }
 
-func GetTableData(sheet string, class int, f *excelize.File) [][]string {
-	startCol := 5
-	endCol := 144
-	timings := [][]string{}
-
-	Days := []string{
-		"Timings",
-		"Monday",
-		"Tuesday",
-		"Wednesday",
-		"Thursday",
-		"Friday",
-	}
-
-	col, err := excelize.ColumnNumberToName(class)
-	if err != nil {
-		panic(err)
-	}
-	timeValue := []string{}
-	for i := 5; i < 32; i += 2 {
-		timeCell := fmt.Sprintf("C%d", i)
-		time, _ := f.GetCellValue(sheet, timeCell)
-		time = strings.ToLower(time)
-		time = strings.ReplaceAll(time, " ", "")
-		timeValue = append(timeValue, time)
-	}
-
-	tempMap := []string{}
-	for i := startCol; i < endCol; i += 2 {
-		timeCell := fmt.Sprintf("C%d", i)
-		time, _ := f.GetCellValue(sheet, timeCell)
-		time = strings.ToLower(time)
-		time = strings.ReplaceAll(time, " ", "")
-
-		var temp string
-		for j := 0; j < 2; j++ {
-			cellId := fmt.Sprintf("%s%d", col, i+j)
-			cell, _ := f.GetCellValue(sheet, cellId)
-			if cell != "" {
-				temp = temp + cell + " "
-			}
-		}
-		if len(temp) == 0 {
-			tempMap = append(tempMap, "Free")
-		} else {
-			tempMap = append(tempMap, temp)
-		}
-		if time == "6:50pm" {
-			timings = append(timings, tempMap)
-			tempMap = []string{}
-		}
-	}
-
-	newtimings := [][]string{}
-	newtimings = append(newtimings, Days)
-	for i := 0; i < 14; i++ {
-		temp := []string{}
-		temp = append(temp, timeValue[i])
-		for _, d := range timings {
-			temp = append(temp, d[i])
-		}
-		newtimings = append(newtimings, temp)
-	}
-
-	return newtimings
-}
+// /^[A-Z]{3}[0-9]{3}/gm
