+++ templates/table.html
@@ -2,7 +2,6 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <!-- Primary Meta Tags -->
     <title>Time Table Generator</title>
     <meta name="title" content="Time Table Generator" />
     <meta
@@ -10,7 +9,6 @@
       content="Quick Time Table Generator for Thapar University Students"
     />
 
-    <!-- Open Graph / Facebook -->
     <meta property="og:type" content="website" />
     <meta property="og:url" content="https://timetable.mlsctiet.com/" />
     <meta property="og:title" content="Time Table Generator" />
@@ -23,7 +21,6 @@
       content="https://www.utkarsh.ninja/favicon-32x32.png"
     />
 
-    <!-- Twitter -->
     <meta property="twitter:card" content="summary_large_image" />
     <meta property="twitter:url" content="https://timetable.mlsctiet.com/" />
     <meta property="twitter:title" content="Time Table Generator" />
