+++ utils/error.go
@@ -1,0 +1,7 @@
+package utils
+
+func HandleError(err error) {
+	if err != nil {
+		panic(err)
+	}
+}
